requirements.txt
streamlit
PyPDF2
scikit-learn
nltk
wordninja
fpdf2
numpy
